Extract shared ticket flow operator
